FIXED - Double books - FIXED - Date picker - actually keep it. FIXED - Quote splash - if none, go to book actions. FIXED - Login - go straight to login. FIXED - On create user - skip login step. FIXED - Password - log straight in - option. FIXED -if you logout - prompt for login on next login. FIXED - book names with a space in them. FIXED - multiline quote fix.

+++ QB/QB/Objects/QBUser.h
@@ -1,7 +1,7 @@
 #import "QB.h"
 
 @interface QBUser : ManagedPropertiesObject <SerializeByDefault>
-@property (nonatomic, assign) int64_t uuid;
+@property (nonatomic, assign) int uuid;
 @property (nonatomic, retain) NSString* email;
 @property (nonatomic, retain) NSString* firstName;
 @property (nonatomic, retain) NSString* lastName;
